mesh: Defunct block extension + bone-accessor signature widening + Mortar::MeshDraw helpers
Round 2 Mesh follow-up per audit (2026-05-04). Closes the 22 remaining
missing-symbol entries by classifying each into Defunct / signature-fix /
new-stub.

Defunct block (Mesh.h) extended with 5 new // Defunct: lines covering
addresses 0x001b08e8 (RebuildEffectBindings), 0x001b10d8 (2-arg ctor),
0x00193ed8 / 0x00193edc / 0x00193ee0 (binary-stub DrawCube/Line/Sphere
file-scope helpers). Now documents 11 Defunct addresses total.

Bone accessor signature widening (4 methods, int -> unsigned long for
binary mangling parity):
- Mesh::SetBones(const BoneBinding*, unsigned long)
- Mesh::GetBoneVertTransform(unsigned long) const
- Mesh::GetBoneWorldTransform(unsigned long) const
- Mesh::GetBoneLocalTransform(unsigned long) const
Caller cascade: 1 site in MeshManager.cpp:300 cast (int) -> (unsigned long).

New file-scope draw helpers (MeshDraw.h/.cpp):
- DrawTris(QUADCUSTOMVERTEX*, count, primType, blend, fx) per binary spec.
- DrawTriList / DrawTriStrip wrappers (primType = GL_TRIANGLES / STRIP).
- DrawQuadUnCached overloads (default 1x1 + sized variant).
- DrawQuad (textured + transformed) wrapper.
- DrawCube / DrawLine / DrawSphere stubs (binary stubs returning input).

Namespace deviation: C++ forbids namespace Mortar::Mesh alongside class
Mortar::Mesh. Helpers placed under namespace Mortar::MeshDraw instead;
ACCEPT-cosmetic mangling deviation documented in triage.json.

Triage entries added (3 ACCEPT-cosmetic):
- Mesh::Draw(_Matrix44<float> const&) -- typedef mangling.
- Mesh::GetBounds() const -- out-ref idiom vs binary's Bounds3D struct return.
- Mesh::BindSkeleton(Skeleton const&) -- ptr vs ref for null-safety.

5 Defunct + 12 // Binary @ + 3 triage entries. Build green for mortar_engine.

+++ src/engine/asset/MeshManager.cpp
@@ -297,7 +297,7 @@ SmartPtr<Model> MeshManager::LoadMeshInternal(const char* path) {
                     loader.ReadBytes(&bones[i].m_BoundsMax, sizeof(Vec3));
                 }
             }
-            mesh->SetBones(bones.data(), (int)boneCount);
+            mesh->SetBones(bones.data(), (unsigned long)boneCount);
         }
 
         // Read<ulong> → materialCount + per-material sub-resource
